admin

+++ app/Http/Controllers/AdminController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use Illuminate\Support\Facades\Validator;
+use App\UE;
 
 class AdminController extends Controller
 {
@@ -31,9 +33,33 @@ class AdminController extends Controller
         if (request('btn-accept') != null) {
              $user -> confirmed = true;
              $user -> save();
-        } else if (request('btn-decline') != null) {
-            $user->delete();
+        } else if (request('btn-decline') != null && !$user->confirmed) {
+               $user->delete();
         }
         return view ('admin.index');
+    }
+
+    public function createUEView() {
+        return view('admin.create-ue');
+    }
+    public function createUE(Request $request) {
+        $data = $request->all();
+
+        $validator = Validator::make($data, [
+            'name' => ['required', 'string', 'max:255'],
+            'annee' => ['required', 'integer','max:5000','min:0']
+        ]);
+        $validator->validate();
+
+        UE::create([
+            'nom' => $data['name'],
+            'annee' => $data['annee']
+        ]);
+
+        return view('admin.create-ue');
+    }
+
+    public function manage() {
+        return view('admin.manage-users');
     }
 }
